requirements.txt
Django==1.10
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-extra-fields==0.9
django-filter==1.0.4
django-markdown-deux==1.0.5
django-pagedown==0.1.3
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
djangorestframework-yaml==1.0.3
Markdown==2.6.9
markdown2==2.3.4
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
PyJWT==1.5.3
PyYAML==3.12
